moved the astype(uint8) on top in load_image

+++ python_01/ex05/load_image.py
@@ -9,9 +9,9 @@ def ft_load(path: str):
 
     :param path: the path to the image.
     """
-    data = img.imread(path)
+    data = img.imread(path).astype(np.uint8)
     print(f"The shape of the image is: {data.shape}")
-    return data.astype(np.uint8)
+    return data
 
 
 __all__ = "ft_load",
